Solve problem 63
How many N-digit integers are there of the form a^N, for any integer a?

An initial insight is to realize that 10^N has N+1 digits, so an upper
bound on a is 10.

The solution is to find the largest and smallest N-digit numbers, and
take their Nth roots. Round up on the smaller root and down on the
larger one. Have an extra check because floating point arithmetic
becomes imprecise with such large numbers.