Solution to problem 65
The sum of the digits of the 100th convergent of e. The nth convergent
is the continued fraction expanded to n terms.

Python provides a fractions module that handles the rational numbers.

Sum has to happen backwards.

Use metaclass to enable iteration and slicing on a top-level class
object just for fun.